refactor: Models (#1177)
