+++ BALSAMIC/utils/metrics.py
@@ -1,0 +1,7 @@
+"""QC metrics utility methods."""
+from BALSAMIC.models.metrics import MetricValidationModel
+
+
+def validate_qc_metrics(metrics: dict) -> dict:
+    """Returns a set of validated QC metrics."""
+    return MetricValidationModel(metrics=metrics).dict()["metrics"]
